fix(#1492): align PR hook receipt body and CI gates
Gate-Record: .workflow/records/1492-adr042-local-gate-receipts.json

Task-Kind: maintenance

Issue: #1492
