+++ scripts/hooks/check-gate-before-pr.sh
@@ -69,12 +69,48 @@ done <<EOF
 $BYPASS_LABELS
 EOF
 
+PR_BODY=$(SCISTUDIO_CMD="$CMD" python - <<'PY'
+import os
+import shlex
+import sys
+from pathlib import Path
+
+try:
+    tokens = shlex.split(os.environ.get("SCISTUDIO_CMD", ""))
+except ValueError:
+    sys.exit(1)
+
+index = 0
+while index < len(tokens):
+    token = tokens[index]
+    value = None
+    if token in {"--body", "-b"} and index + 1 < len(tokens):
+        value = tokens[index + 1]
+        index += 1
+    elif token.startswith("--body=") or token.startswith("-b="):
+        value = token.split("=", 1)[1]
+    elif token == "--body-file" and index + 1 < len(tokens):
+        value = Path(tokens[index + 1]).read_text(encoding="utf-8")
+        index += 1
+    elif token.startswith("--body-file="):
+        value = Path(token.split("=", 1)[1]).read_text(encoding="utf-8")
+    if value is not None:
+        print(value, end="")
+        sys.exit(0)
+    index += 1
+sys.exit(2)
+PY
+) || {
+  echo '{"hookSpecificOutput":{"hookEventName":"PreToolUse","permissionDecision":"deny","permissionDecisionReason":"ADR-042 gate: gh pr create must provide the real PR body via --body or --body-file. Use scripts/scistudio_pr_create.py for receipt-backed PR creation."}}'
+  exit 0
+}
+
 if [ -n "$BROAD_BYPASS" ]; then
   OUTPUT=$(PYTHONPATH=src python -m scistudio.qa.governance.gate_record pr-ready \
     --repo-root . \
     --base "$BASE_REF" \
     --head HEAD \
-    --pr-body "$CMD" \
+    --pr-body "$PR_BODY" \
     "${LABEL_ARGS[@]}" 2>&1) || {
     REASON=$(printf '%s' "$OUTPUT" | python -c "import json,sys; print(json.dumps(sys.stdin.read()[:1800]))")
     echo "{\"hookSpecificOutput\":{\"hookEventName\":\"PreToolUse\",\"permissionDecision\":\"deny\",\"permissionDecisionReason\":$REASON}}"
@@ -84,7 +120,7 @@ if [ -n "$BROAD_BYPASS" ]; then
   exit 0
 fi
 
-if ! echo "$CMD" | grep -qiE '(closes|fixes|resolves)[[:space:]]+#?[0-9]+'; then
+if ! printf '%s' "$PR_BODY" | grep -qiE '(closes|fixes|resolves)[[:space:]]+#?[0-9]+'; then
   echo '{"hookSpecificOutput":{"hookEventName":"PreToolUse","permissionDecision":"deny","permissionDecisionReason":"ADR-042 gate: gh pr create command must include a PR body with Closes/Fixes/Resolves #N."}}'
   exit 0
 fi
@@ -93,7 +129,7 @@ OUTPUT=$(PYTHONPATH=src python -m scistudio.qa.governance.gate_record pr-ready \
   --repo-root . \
   --base "$BASE_REF" \
   --head HEAD \
-  --pr-body "$CMD" \
+  --pr-body "$PR_BODY" \
   "${LABEL_ARGS[@]}" 2>&1) || {
   REASON=$(printf '%s' "$OUTPUT" | python -c "import json,sys; print(json.dumps(sys.stdin.read()[:1800]))")
   echo "{\"hookSpecificOutput\":{\"hookEventName\":\"PreToolUse\",\"permissionDecision\":\"deny\",\"permissionDecisionReason\":$REASON}}"
@@ -128,7 +164,7 @@ RECEIPT_OUTPUT=$(PYTHONPATH=src python -m scistudio.qa.governance.gate_receipt v
   --gate-record "$GATE_RECORD" \
   --base "$BASE_REF" \
   --head HEAD \
-  --pr-body "$CMD" 2>&1) || {
+  --pr-body "$PR_BODY" 2>&1) || {
   REASON=$(printf '%s' "$RECEIPT_OUTPUT" | python -c "import json,sys; print(json.dumps(sys.stdin.read()[:1800]))")
   echo "{\"hookSpecificOutput\":{\"hookEventName\":\"PreToolUse\",\"permissionDecision\":\"deny\",\"permissionDecisionReason\":$REASON}}"
   exit 0
